initial commit: sloppy
this is the state of the code on my laptop as of now. its got a bunch of commented out stuff and a bunch of haphazard stuff i need to edit and delete and fix.

this is a drag and drop of my src file. Im not sure that Im doing this right... This is kind of an experiment.